+++ ostap/stats/corr2d.py
@@ -6,21 +6,23 @@
 #  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
 #  @date   2014-06-08
 # =============================================================================
-"""Simple 2D-decorrelation transformation """
+""" Simple 2D-decorrelation transformation """
 # =============================================================================
 __version__ = "$Revision$"
 __author__  = "Vanya BELYAEV Ivan.Belyaev@itep.ru"
 __date__    = "2014-06-08"
 __all__     = ( 'Corr2D', )
 # =============================================================================
-from   ostap.core.core import cpp , WSE
+from   ostap.core.core    import cpp , WSE, Ostap 
+from   ostap.trees.cuts   import expression_types 
+import ostap.math.linalg
 import ROOT,math
 # =============================================================================
 # logging 
 # =============================================================================
 from   ostap.logger.logger import getLogger
-if '__main__' ==  __name__ : logger = getLogger( 'Ostap.Corr2D' )
-else                       : logger = getLogger( __name__ )
+if '__main__' ==  __name__ : logger = getLogger ( 'ostap.stats.corr2D' )
+else                       : logger = getLogger ( __name__ )
 # =============================================================================
 ## get error-function 
 if  not hasattr ( math , 'erf' ) :
@@ -31,6 +33,7 @@ if  not hasattr ( math , 'erf' ) :
 ## error function
 #  @see http://en.wikipedia.org/wiki/Error_function
 erf = math.erf
+# ============================================================================
 ## get Gaussian cdf 
 if not hasattr ( math , 'gauss_cdf' ) :
     from ostap.math.math_ve import gauss_cdf as _gauss_cdf
@@ -39,12 +42,17 @@ if not hasattr ( math , 'gauss_cdf' ) :
 # =============================================================================
 gauss_cdf = math.gauss_cdf
 # =============================================================================
-## simple 2D-decorrelation transformation 
+## invoke decorations of  matrices & vectors 
+C2 = Ostap.Math.SymMatrix ( 2 )
+V2 = Ostap.Math.Vector    ( 2 )
+# =============================================================================
+## Simple 2D-decorrelation transformation 
 #  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
 #  @date   2014-06-08
 class Corr2D(object) :
-    """Simple 2D-decorrelation transformation 
+    """ Simple 2D-decorrelation transformation 
     """
+    # =========================================================================
     ## constructor
     #  @param chain     (INPUT) tree/chain or dataset 
     #  @param var1      (INPUT) expression for the first variable 
@@ -52,165 +60,253 @@ class Corr2D(object) :
     #  @param selection (INPUT) cuts, if needed
     #  @param first     (INPUT) the first event
     #  @param last      (INPUT) the last event     
-    def __init__  ( self              ,
-                    chain             ,
-                    var1              ,
-                    var2              , 
-                    selection = None  ,
-                    first     = 0     ,
-                    last      = 2**63 ) :
+    def __init__  ( self           ,
+                    dataset        ,
+                    exprs1         ,
+                    exprs2         ,
+                    selection = '' , *args )  :
 
-        import ostap.math.linalg
+        ## extra arguments 
+        self.__args = args
         
-        SV    =  Ostap.StatVar
-        ##
-        self.stat1 = WSE () 
-        self.stat2 = WSE ()
-        self.cov2  = Ostap.SymMatrix(2) ()
-        ##
-        self.var1 = var1
-        self.var2 = var2
+        assert exprs1 and isinstance ( exprs1 , expression_types ) , \
+            "Invalid type of `exprs1`:%s" % type ( exprs1 )
+        assert exprs2 and isinstance ( exprs2 , expression_types ) , \
+            "Invalid type of `exprs2`:%s" % type ( exprs2 )
+        assert not selection or isinstance ( selection , expression_types ) , \
+            "Invalid type of `selection`: %s" % type ( selection )
         
+        self.__var1      = str ( exprs1    ).strip () 
+        self.__var2      = str ( exprs2    ).strip () 
+        self.__selection = str ( selection ).strip () if selection else ''
+
+        ## get the statistics&covariances 
+        self.__wcov = Ostap.StatVar.statCov ( dataset        ,
+                                              self.var1      ,
+                                              self.var2      ,
+                                              self.selection ,
+                                              *self.args     ) 
+        
+        self.__counter1 = self.__wcov.counter1 ()  
+        self.__counter2 = self.__wcov.counter2 ()
+
+        ## the covariance matrix 
+        self.__cov2     = Ostap.Math.covariance  (  self.__wcov )
+        
+        ## the correlation matrix 
+        self.__corr     = Ostap.Math.correlation (  self.__wcov )
+
+        ## eigensystem: eigenvalues and transformation 
+        self.__evalues, self.__evectors = self.__cov2.eigenVectors( sorted = True , ascending = False )
+
+        ## the first and second eigenvectors
+        self.__evector1 = self.__evectors.column ( 0 ) 
+        self.__evector2 = self.__evectors.column ( 1 ) 
+        
+        ## get the first an second rows of transformation matrix
+        self.__row1 = self.__evectors.row ( 0 ) 
+        self.__row2 = self.__evectors.row ( 1 )
+        
+        ## global bias 
+        self.__delta = V2()
+        self.__delta [ 0 ] = float ( self.__counter1.mean() ) 
+        self.__delta [ 1 ] = float ( self.__counter2.mean() ) 
+
+        nmean1 = -1 * self.__delta [ 0 ] 
+        nmean2 = -1 * self.__delta [ 1 ] 
+        
+        dv1 = '((%s)%+.16g)' % ( self.var1 , nmean1 )
+        dv2 = '((%s)%+.16g)' % ( self.var2 , nmean2 )
+
+        self.__decorrelated1 = "(%+.16g*%s%+.16g*%s)" % ( self.__row1 [ 0 ] , dv1 , self.__row1 [ 1 ] , dv2 )
+        self.__decorrelated2 = "(%+.16g*%s%+.16g*%s)" % ( self.__row2 [ 0 ] , dv1 , self.__row2 [ 1 ] , dv2 )
+        
+        scale1 = math.sqrt ( self.__evalues [ 0 ] )
+        scale2 = math.sqrt ( self.__evalues [ 1 ] )
+
+        ## get the first and second rows of transformation matrix and scale them 
+        self.__srow1 = self.__row1 / scale1 
+        self.__srow2 = self.__row2 / scale2 
+
+        self.__normalized1 = "(%+.16g*%s%+.16g*%s)" % ( self.__srow1 [ 0 ] , dv1 , self.__srow1 [ 1 ] , dv2 )
+        self.__normalized2 = "(%+.16g*%s%+.16g*%s)" % ( self.__srow2 [ 0 ] , dv1 , self.__srow2 [ 1 ] , dv2 )
+        
+        ## get the first and second rows of transformation matrix and scale them
+        sqrt2 = math.sqrt ( 2.0 )
+        self.__ssrow1 = self.__srow1  / sqrt2 
+        self.__ssrow2 = self.__srow2  / sqrt2 
+        
+        self.__u1 = "(%+.16g*%s%+.16g*%s)" % ( self.__ssrow1 [ 0 ] , dv1 , self.__ssrow1 [ 1 ] , dv2 )
+        self.__u2 = "(%+.16g*%s%+.16g*%s)" % ( self.__ssrow2 [ 0 ] , dv1 , self.__ssrow2 [ 1 ] , dv2 )
+        
+        self.__uniform1 = "(0.5+0.5*TMath::Erf(%s))" % self.__u1 
+        self.__uniform2 = "(0.5+0.5*TMath::Erf(%s))" % self.__u2
+
+
+    @property
+    def var1 ( self ) :
+        """`var1' : the first expression, same as `expr1`"""
+        return self.__var1
+
+    @property
+    def var2 ( self ) :
+        """`var2' : the second expression, same as `expr2`"""
+        return self.__var2
+
+    @property
+    def expr1 ( self ) :
+        """`expr1' : the first expression, same as `var1`"""
+        return self.var1
+    
+    @property
+    def expr2 ( self ) :
+        """`expr2' : the second expression, same as `var2`"""
+        return self.var2
+
+    @property
+    def selection ( self ) :
+        """`selection` : selection/weight  expression (same as `cuts` or `weight`)"""
+        return self.__selection
+
+    @property
+    def cuts ( self ) :
+        """`cuts` : selection/weight  expression (same as `selection` or `weight`)"""
+        return self.selection
+
+    @property
+    def weight ( self ) :
+        """`weight` : selection/weight  expression (same as `selection` or `cuts`)"""
+        return self.selection
+
+    @property
+    def args  ( self ) :
+        """`args` : extra arguments for `Ostap.StatVar.statCov` call"""
+        return self.__args
+    
+    @property
+    def covariance ( self ) :
+        """`covariance' :  `Ostap::Math::(W)Covariance` object"""
+        return self.__wcov
+
+    @property
+    def counter1 ( self ) : 
+        """`conter1' : counetr for the first expression"""
+        return self.covariance.counter1()
+
+    @property
+    def counter2 ( self ) : 
+        """`conter2' : counetr for the second expression"""
+        return self.covariance.counter2()
+
+    @property
+    def correlation ( self ) :
+        """`correlation` : get the correlation coefficient """
+        return self.covariance.correlation()
+        
+    @property
+    def cov2 ( self ) :
+        """`cov2` : covariance matrix"""
+        return self.__cov2 
+
+    @property
+    def corr ( self ) :
+        """`corr` : correlation matrix"""
+        return self.__corr 
+
+    @property
+    def eigenvalues  ( self ) :
+        """`eigenvalues` : eigenvalues of covariance matrix"""
+        return self.__evalues
+
+    @property
+    def eigenvectors ( self ) :
+        """`eigenvectors` : matrix of eigenvectors: each coluimn is eigenvector, each eigenvector is normalized to 1"""
+        return self.__evectors 
+
+    @property
+    def decorrelated1 ( self ) :
+        """`decorrelated1` : the first decorrelated variable  """
+        return self.__decorrelated1
+        
+    @property
+    def decorrelated2 ( self ) :
+        """`decorrelated2` : the second decorrelated variable  """
+        return self.__decorrelated2
+    
+    @property
+    def normalized1 ( self ) :
+        """`normalized1` : the first decorrelated&normalized variable  """
+        return self.__normalized1
+
+    @property
+    def normalized2 ( self ) :
+        """`normalized2` : the second decorrelated&normalized variable  """
+        return self.__normalized2
+
+    @property
+    def uniform1 ( self ) :
+        """`uniform1` : the first decorrelated/normalized&uniform variable  """
+        return self.__uniform1
+    
+    @property
+    def uniform2 ( self ) :
+        """`uniform2` : the second decorrelated/normalized&uniform variable  """
+        return self.__uniform2
+
+    # =========================================================================
+    ## Make transformation from original to decorrelated variables
+    #  @code
+    #  x  , y  = ...
+    #  xd , yd = corr2d.decorrelated ( x , y ) 
+    #  @endcode 
+    def decorrelated  ( self , x , y ) :
+        """ Make a transformation from original to decorrelated variables
+        >>> x  , y  = ...
+        >>> xd , yd = corr2d.decorrelated ( x , y ) 
+        """
+        dv = V2 ( x , y ) - self.__delta
         ## 
-        if isinstance ( chain , ROOT.RooDataSet ) :
-            chain = chain.store().tree() 
+        xd = self.__row1 * dv
+        yd = self.__row2 * dv
         ## 
-        if selection :
-            self.num   = SV.statCov ( chain      ,
-                                      var1       ,
-                                      var2       ,
-                                      selection  ,
-                                      self.stat1 ,
-                                      self.stat2 ,                                 
-                                      self.cov2  ,
-                                      first      , last ) 
-        else :
-            self.num   = SV.statCov ( chain      ,
-                                      var1       ,
-                                      var2       ,
-                                      self.stat1 ,
-                                      self.stat2 ,                                 
-                                      self.cov2  ,
-                                      first      , last ) 
+        return xd , yd 
 
-
-        self.corr = Ostap.SymMatrix(2)()
-
-        rms1 = self.stat1.rms()
-        rms2 = self.stat2.rms()
-        
-        self.corr[0,0] = self.cov2(0,0) / ( rms1 * rms1 )
-        self.corr[0,1] = self.cov2(0,1) / ( rms1 * rms2 )
-        self.corr[1,1] = self.cov2(1,1) / ( rms2 * rms2 )
-        
-        self.correlation = self.corr(0,1)
-        
-        ## get eigen vectors 
-        self.e = self.cov2.eigenVectors()
-        
-        self.vct0  = self.e[1][0]
-        self.vct1  = self.e[1][1]
-
-        import math
-
-        ## normalized eigen vectors 
-        self.vct0 /= math.sqrt ( self.e[0][0] ) 
-        self.vct1 /= math.sqrt ( self.e[0][1] )
-        
-        self.m1 = -1*self.stat1.mean().value()
-        self.m2 = -1*self.stat2.mean().value()
-        
-        self.nvar1 = "(%+g*((%s)%+g)%+g*((%s)%+g))" % ( self.vct0[0] , self.var1 , self.m1 , self.vct0[1] , self.var2 , self.m2 ) 
-        self.nvar2 = "(%+g*((%s)%+g)%+g*((%s)%+g))" % ( self.vct1[0] , self.var1 , self.m1 , self.vct1[1] , self.var2 , self.m2 ) 
-
-        logger.info ( 'Correlation %.3f%%' % ( 100 * self.correlation )  )
-        
-        logger.info ( 'The 1st decorrelated variable:\n %s ' % self.nvar1 )
-        logger.info ( 'The 2nd decorrelated variable:\n %s ' % self.nvar2 )
-
-        ## normalize eigenvectors for 1/sqrt(2), just for convinency of erf. 
-        sqr2i         = math.sqrt(0.5) 
-        self.nvct0    = Ostap.Vector2() 
-        self.nvct1    = Ostap.Vector2()
-        
-        self.nvct0[0] = self.vct0[0]*sqr2i
-        self.nvct0[1] = self.vct0[1]*sqr2i
-        self.nvct1[0] = self.vct1[0]*sqr2i
-        self.nvct1[1] = self.vct1[1]*sqr2i
-
-        self.qvar1 = "0.5+0.5*TMath::Erf( %+g*((%s)%+g) %+g*((%s)%+g) )" % ( self.nvct0[0] , self.var1 , self.m1 , self.nvct0[1] , self.var2 , self.m2 ) 
-        self.qvar2 = "0.5+0.5*TMath::Erf( %+g*((%s)%+g) %+g*((%s)%+g) )" % ( self.nvct1[0] , self.var1 , self.m1 , self.nvct1[1] , self.var2 , self.m2 ) 
-
-        logger.info ( 'The 1st decorrelated normalized variable:\n %s ' % self.qvar1 )
-        logger.info ( 'The 2nd decorrelated normalized variable:\n %s ' % self.qvar2 )
-
+    # =========================================================================
+    ## Make transformation from original to decorrelated&normalized variables
+    #  @code
+    #  x  , y  = ...
+    #  xn , yn = corr2d.normalized ( x , y )
+    #  @endcode 
+    def normalized  ( self , x , y ) :
+        """ Make a transformation from original to decorrelated variables
+        >>> x  , y  = ...
+        >>> xn , yn = corr2d.normalized ( x , y )
+        """
+        dv = V2 ( x , y ) - self.__delta
+        ## 
+        xn = self.__srow1 * dv
+        yn = self.__srow2 * dv
+        ## 
+        return xn , yn
     
-    def   fvar1   ( self ) :
-        #
-        v1 = str ( self.qvar1 )
-        v1 = v1.replace ( '0.5+0.5*TMath::Erf' , 'gauss_cdf' )
-        v1 = v1.replace ( self.var1    ,  's.' + self.var1   )
-        v1 = v1.replace ( self.var2    ,  's.' + self.var2   )
-        #
-        return eval ( 'lambda s: ' + v1 )
+    # =========================================================================
+    ## Make transformation from original to decorrelated,normalized&uniform variables
+    #  @code
+    #  x  , y  = ...
+    #  xn , yn = corr2d.uniform ( x , y )
+    #  @endcode 
+    def uniform ( self , x , y ) :
+        """ Make a transformation from original to decorrelated variables
+        >>> x  , y  = ...
+        >>> xn , yn = corr2d.uniform ( x , y (   
+        """
+        dv = V2 ( x , y ) - self.__delta
+        ## 
+        xu = gauss_cdf ( self.__srow1 * dv ) 
+        yu = gauss_cdf ( self.__srow2 * dv ) 
+        ## 
+        return xu , yu
     
-    def   fvar2   ( self ) :
-        #
-        v2 = str ( self.qvar2 )
-        v2 = v2.replace ( '0.5+0.5*TMath::Erf' , 'gauss_cdf' )
-        v2 = v2.replace ( self.var1    ,  's.' + self.var1   )
-        v2 = v2.replace ( self.var2    ,  's.' + self.var2   )
-        #
-        return eval ( 'lambda s: ' + v2 )
-    
-    def   fun2D  ( self ) :
-        #
-        v1 = str ( self.qvar1 )
-        v1 = v1.replace ( '0.5+0.5*TMath::Erf' , 'gauss_cdf' )
-        v1 = v1.replace ( self.var1    ,  's.' + self.var1   )
-        v1 = v1.replace ( self.var2    ,  's.' + self.var2   )
-        #
-        v2 = str ( self.qvar2 )
-        v2 = v2.replace ( '0.5+0.5*TMath::Erf' , 'gauss_cdf' )
-        v2 = v2.replace ( self.var1    ,  's.' + self.var1   )
-        v2 = v2.replace ( self.var2    ,  's.' + self.var2   )
-        #        #
-        return eval ( 'lambda s: ( ' + v1 + ',' + v2 + ' ) ' )
-    
-    def __repr__ ( self ) :
-
-        result  =   'Events                       %s'       % self.num
-        result += '\nStat(var1) [mean/RMS/minmax] %s/%s/%s' % ( self.stat1.mean   () ,
-                                                                self.stat1.rms    () ,
-                                                                self.stat1.minmax () ) 
-        result += '\nStat(var2) [mean/RMS/minmax] %s/%s/%s' % ( self.stat2.mean   () ,
-                                                                self.stat2.rms    () ,
-                                                                self.stat2.minmax () ) 
-        result += '\nCovariance  \n%s' % self.cov2 
-        
-        result += '\nStat(var1)           %s' % self.stat1
-        result += '\nStat(var1) [values ] %s' % self.stat1.values  () 
-        result += '\nStat(var1) [weights] %s' % self.stat1.weights ()
-        result += '\nStat(var2)           %s' % self.stat1
-        result += '\nStat(var2) [values ] %s' % self.stat1.values  () 
-        result += '\nStat(var2) [weights] %s' % self.stat1.weights ()
-        
-        result += '\nCorrelation: %.3f%%\n%s' % ( self.corr(0,1)*100 , self.corr )
-        
-        result += '\nThe first  eigenvector %s '                      % self.vct0
-        result += '\nThe second eigenvector %s '                      % self.vct1
-
-        result += '\nThe 1st decorrelated variable:\n %s '            % self.nvar1 
-        result += '\nThe 2nd decorrelated variable:\n %s '            % self.nvar2 
-        
-        result += '\nThe 1st decorrelated normalized variable:\n %s ' % self.qvar1 
-        result += '\nThe 2nd decorrelated normalized variable:\n %s ' % self.qvar2 
-
-        return result
-
-    __str__ = __repr__ 
-
-
 # =============================================================================
 if '__main__' == __name__ :
         
@@ -218,5 +314,5 @@ if '__main__' == __name__ :
     docme ( __name__ , logger = logger )
 
 # =============================================================================
-# The END 
+##                                                                     The END 
 # =============================================================================